+++ examples/webpack/public/index.js
@@ -15,43 +15,20 @@
 
 'use strict';
 
-// make `.jsx` file requirable by node
-require('node-jsx').install();
+var Routes = require('./routes.jsx');
+var Client = require('react-engine/lib/client');
 
-var express = require('express');
-var renderer = require('react-engine');
+// boot options
+var options = {
+  routes: Routes,
 
-var app = express();
+  // supply a function that can be called
+  // to resolve the file that was rendered.
+  viewResolver: function(viewName) {
+    return require('./views/' + viewName);
+  }
+};
 
-// create the view engine with `react-engine`
-var engine = renderer.server.create();
-
-// set the engine
-app.engine('.jsx', engine);
-
-// set the view directory
-app.set('views', __dirname + '/public/views');
-
-// set jsx as the view engine
-app.set('view engine', 'jsx');
-
-// finally, set the custom view
-app.set('view', renderer.expressView);
-
-//expose public folder as static assets
-app.use(express.static(__dirname + '/public'));
-
-app.use('/', function(req, res) {
-  res.render('index', {
-    title: 'React Engine Express Sample App',
-    name: 'Jordan'
-  });
-});
-
-var server = app.listen(3000, function() {
-
-  var host = server.address().address;
-  var port = server.address().port;
-
-  console.log('Example app listening at http://%s:%s', host, port);
+document.addEventListener('DOMContentLoaded', function onLoad() {
+  Client.boot(options);
 });